requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.27.1
